52 fix scaling physics (#54)
Fix scaling physics to some extent by having the level scale instead of the ball

Although this does not allow for much more zooming in (now scaling up) as I had hoped, it does provide several benefits and improvements:

1. It fixes issues with the restitution not pushing the ball out of the floor at smaller scales
2. It makes the scaling slightly more stable, allowing slightly smaller scales (though not by much) without much physics weirdness
3. It allows much larger scales without the background going funky
4. The background now stays static wrt scale (unintended, but I like it as we don't yet have the background using the new shader)
5. The scaling code is now separate, making for smaller and easier to digest chunks of code

The only downside is that there seems to be some weird issue where the scaling oddly causes the ball to slow down abruptly. This only occurs for some levels / slopes though, and I'm not sure it's directly caused by / related to the scaling. It will need some investigation, but is relatively minor.